first pass at unionizing the blooms
