+++ probables/blooms/bloom.py
@@ -11,13 +11,13 @@ import os
 import mmap
 from struct import (pack, unpack, calcsize, Struct)
 from shutil import (copyfile)
-from binascii import (hexlify, unhexlify)
+from . basebloom import (BaseBloom)
 from .. exceptions import (InitializationError, NotSupportedError)
 from .. hashes import (default_fnv_1a)
 from .. utilities import (is_hex_string, is_valid_file)
 
 
-class BloomFilter(object):
+class BloomFilter(BaseBloom):
     ''' Simple Bloom Filter implementation for use in python;
         It can read and write the same format as the c version
         (https://github.com/barrust/bloom)
@@ -41,84 +41,9 @@ class BloomFilter(object):
     def __init__(self, est_elements=None, false_positive_rate=None,
                  filepath=None, hex_string=None, hash_function=None):
         ''' setup the basic values needed '''
-        self._bloom = None
-        self.__num_bits = 0  # number of bits
-        self.__est_elements = 0
-        self.__fpr = 0.0
-        self.__number_hashes = 0
-        self.__bloom_length = self.number_bits // 8
-        self.__hash_func = default_fnv_1a
-        self._els_added = 0
-        self._on_disk = False  # not on disk
-
-        if is_valid_file(filepath):
-            self.__load(filepath, hash_function)
-        elif is_hex_string(hex_string):
-            self._load_hex(hex_string, hash_function)
-        elif est_elements is not None and false_positive_rate is not None:
-            self._set_optimized_params(est_elements, false_positive_rate, 0,
-                                       hash_function)
-            self._bloom = [0] * self.bloom_length
-        else:
-            msg = ('Insufecient parameters to set up the Bloom Filter')
-            raise InitializationError(msg)
-
-    @property
-    def false_positive_rate(self):
-        ''' float: The maximum desired false positive rate
-
-            Note:
-                Not settable '''
-        return self.__fpr
-
-    @property
-    def estimated_elements(self):
-        ''' int: The maximum number of elements estimated to be added at setup
-
-            Note:
-                Not settable '''
-        return self.__est_elements
-
-    @property
-    def number_hashes(self):
-        ''' int: The number of hashes required for the Bloom Filter hashing
-            strategy
-
-            Note:
-                Not settable '''
-        return self.__number_hashes
-
-    @property
-    def number_bits(self):
-        ''' int: Number of bits in the Bloom Filter
-
-            Note:
-                Not settable '''
-        return self.__num_bits
-
-    @property
-    def elements_added(self):
-        ''' int: Number of elements added to the Bloom Filter
-
-        Note:
-            Not settable '''
-        return self._els_added
-
-    @property
-    def is_on_disk(self):
-        ''' bool: Is the Bloom Filter on Disk or not
-
-        Note:
-            Not settable '''
-        return self._on_disk
-
-    @property
-    def bloom_length(self):
-        ''' int: Length of the Bloom Filter array
-
-        Note:
-            Not settable '''
-        return self.__bloom_length
+        super(BloomFilter, self).__init__('regular', est_elements,
+                                          false_positive_rate, filepath,
+                                          hex_string, hash_function)
 
     def __str__(self):
         ''' correctly handle python 3 vs python2 encoding if necessary '''
@@ -146,29 +71,6 @@ class BloomFilter(object):
                             self.current_false_positive_rate(),
                             self.export_size(), self.__cnt_number_bits_set(),
                             on_disk)
-
-    def __contains__(self, key):
-        ''' setup the `in` keyword '''
-        return self.check(key)
-
-    def clear(self):
-        ''' Clear or reset the Bloom Filter '''
-        self._els_added = 0
-        for idx in range(self.bloom_length):
-            self._bloom[idx] = self._get_set_element(0)
-
-    def hashes(self, key, depth=None):
-        ''' Return the hashes based on the provided key
-
-            Args:
-                key (str): Description of arg1
-                depth (int): Number of permutations of the hash to generate; \
-                if None, generate `number_hashes`
-            Returns:
-                List(int): A list of the hashes for the key in int form
-        '''
-        tmp = depth if depth is not None else self.number_hashes
-        return self.__hash_func(key, tmp)
 
     def add(self, key):
         ''' Add the key to the Bloom Filter
@@ -239,7 +141,7 @@ class BloomFilter(object):
         if self.__verify_bloom_similarity(second) is False:
             return None
         res = BloomFilter(self.estimated_elements, self.false_positive_rate,
-                          hash_function=self.__hash_func)
+                          hash_function=self.hash_function)
 
         for i in list(range(0, self.bloom_length)):
             res._bloom[i] = self._get_element(i) & second._get_element(i)
@@ -264,7 +166,7 @@ class BloomFilter(object):
         if self.__verify_bloom_similarity(second) is False:
             return None
         res = BloomFilter(self.estimated_elements, self.false_positive_rate,
-                          hash_function=self.__hash_func)
+                          hash_function=self.hash_function)
 
         for i in list(range(0, self.bloom_length)):
             res._bloom[i] = self._get_element(i) | second._get_element(i)
@@ -297,70 +199,6 @@ class BloomFilter(object):
             return 1.0
         return count_int / count_union
 
-    def export(self, filename):
-        ''' Export the Bloom Filter to disk
-
-            Args:
-                filename (str): The filename to which the Bloom Filter will \
-                be written.
-        '''
-        with open(filename, 'wb') as filepointer:
-            rep = 'B' * self.bloom_length
-            filepointer.write(pack(rep, *self._bloom))
-            filepointer.write(pack('QQf', self.estimated_elements,
-                                   self.elements_added,
-                                   self.false_positive_rate))
-
-    def __load(self, filename, hash_function=None):
-        ''' load the Bloom Filter from file '''
-        # read in the needed information, and then call _set_optimized_params
-        # to set everything correctly
-        with open(filename, 'rb') as filepointer:
-            offset = calcsize('QQf')
-            filepointer.seek(offset * -1, os.SEEK_END)
-            mybytes = unpack('QQf', filepointer.read(offset))
-            self._set_optimized_params(mybytes[0], mybytes[2],
-                                       mybytes[1], hash_function)
-
-            # now read in the bit array!
-            filepointer.seek(0, os.SEEK_SET)
-            offset = calcsize('B') * self.bloom_length
-            rep = 'B' * self.bloom_length
-            self._bloom = list(unpack(rep, filepointer.read(offset)))
-
-    def export_hex(self):
-        ''' Export the Bloom Filter as a hex string
-
-            Return:
-                str: Hex representation of the Bloom Filter
-        '''
-        mybytes = pack('>QQf', self.estimated_elements,
-                       self.elements_added, self.false_positive_rate)
-        bytes_string = hexlify(bytearray(self._bloom)) + hexlify(mybytes)
-        if sys.version_info > (3, 0):  # python 3 gives us bytes
-            return str(bytes_string, 'utf-8')
-        return bytes_string
-
-    def _load_hex(self, hex_string, hash_function=None):
-        ''' placeholder for loading from hex string '''
-        offset = calcsize('>QQf') * 2
-        stct = Struct('>QQf')
-        tmp_data = stct.unpack_from(unhexlify(hex_string[-offset:]))
-        self._set_optimized_params(tmp_data[0], tmp_data[2], tmp_data[1],
-                                   hash_function)
-        tmp_bloom = unhexlify(hex_string[:-offset])
-        rep = 'B' * self.bloom_length
-        self._bloom = list(unpack(rep, tmp_bloom))
-
-    def export_size(self):
-        ''' Calculate the size of the bloom on disk
-
-            Returns:
-                int: Size of the Bloom Filter when exported to disk
-        '''
-        tmp_b = calcsize('B')
-        return (self.bloom_length * tmp_b) + calcsize('QQf')
-
     def estimate_elements(self):
         ''' Estimate the number of elements added
 
@@ -382,25 +220,6 @@ class BloomFilter(object):
         dbl = num / float(self.number_bits)
         exp = math.exp(dbl)
         return math.pow((1 - exp), self.number_hashes)
-
-    def _set_optimized_params(self, estimated_elements, false_positive_rate,
-                              elements_added, hash_function):
-        ''' set the parameters to the optimal sizes '''
-        if hash_function is None:
-            self.__hash_func = default_fnv_1a
-        else:
-            self.__hash_func = hash_function
-        self.__est_elements = estimated_elements
-        fpr = pack('f', float(false_positive_rate))
-        self.__fpr = unpack('f', fpr)[0]  # to mimic the c version!
-        self._els_added = elements_added
-        # optimal caluclations
-        n_els = self.estimated_elements
-        fpr = float(self.__fpr)
-        m_bt = math.ceil((-n_els * math.log(fpr)) / 0.4804530139182)  # ln(2)^2
-        self.__number_hashes = int(round(math.log(2.0) * m_bt / n_els))
-        self.__num_bits = int(m_bt)
-        self.__bloom_length = int(math.ceil(m_bt / (8 * 1.0)))
 
     def __verify_bloom_similarity(self, second):
         ''' can the blooms be used in intersection, union, or jaccard index '''
@@ -481,9 +300,13 @@ class BloomFilterOnDisk(BloomFilter):
         if est_elements is not None and false_positive_rate is not None:
             # no need to check the file since this will over write it
             fpr = false_positive_rate
-            super(BloomFilterOnDisk,
-                  self)._set_optimized_params(est_elements, fpr, 0,
-                                              hash_function)
+            vals = super(BloomFilterOnDisk,
+                         self)._set_optimized_params(est_elements, fpr, 0,
+                                                     hash_function)
+            self.__hash_func = vals[0]
+            self.__fpr = vals[1]
+            self.__number_hashes = vals[2]
+            self.__num_bits = vals[3]
             # do the on disk things
             with open(filepath, 'wb') as filepointer:
                 for _ in range(self.bloom_length):
@@ -520,12 +343,19 @@ class BloomFilterOnDisk(BloomFilter):
             offset = calcsize('QQf')
             filepointer.seek(offset * -1, os.SEEK_END)
             mybytes = unpack('QQf', filepointer.read(offset))
-            self._set_optimized_params(mybytes[0], mybytes[2], mybytes[1],
-                                       hash_function)
+            vals = super(BloomFilterOnDisk,
+                         self)._set_optimized_params(mybytes[0], mybytes[2],
+                                                     mybytes[1], hash_function)
+        self.__hash_func = vals[0]
+        self.__fpr = vals[1]
+        self.__number_hashes = vals[2]
+        self.__num_bits = vals[3]
         self.__file_pointer = open(filepath, 'r+b')
         self._bloom = mmap.mmap(self.__file_pointer.fileno(), 0)
         self._on_disk = True
         self.__filename = filepath
+        print(self.false_positive_rate, self.number_hashes, self.number_bits)
+        print(vals[1:])
 
     def export(self, filename):
         ''' Export to disk if a different location
